requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
idna==2.10
lxml==4.6.2
multidict==6.0.2
pycparser==2.21
PyNaCl==1.4.0
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.25.1
requests-futures==1.0.0
six==1.16.0
soupsieve==2.1
stem==1.8.0
torrequest==0.1.0
typing_extensions==4.1.1
urllib3==1.26.2
yarl==1.7.2
